Tests added and few changes are done
