+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start":"node index.js"
+    "start": "node index.js",
+    "test": "node --experimental-vm-modules node_modules/.bin/jest --coverage"
   },
   "author": "",
   "license": "ISC",
@@ -22,6 +23,12 @@
     "socket.io": "^4.5.2"
   },
   "devDependencies": {
-    "prisma": "^4.3.1"
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "jest": "^29.1.1",
+    "jest-mock-extended": "^3.0.1",
+    "prisma": "^4.3.1",
+    "superagent": "^8.0.0",
+    "supertest": "^6.2.4"
   }
 }
