+++ sockets.js
@@ -1,0 +1,3 @@
+const sockets = [];
+
+module.exports = sockets;
